### 2022-07-20 #4573 @pdebuyl
Hi @arfon . I could, but not before mid-august. Seems a bit more GIS than what I do though.
