### 2022-08-10 #4664 @editorialbot
[REVIEW]: ur-scape: harnessing data for stakeholder participation in city-making processes
**Submitting author:** @luyuhao0326 (<a href="http://orcid.org/0000-0001-8574-1999">Yuhao Lu</a>)
**Repository:** https://github.com/UrbanRuralSystems/ur-scape
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.9.96
**Archive:** 10.3929/ethz-b-000578490
