### 2022-09-09 UrbanRuralSystems/ur-scape#8 @winstonyym
JOSS Review – Software paper
[#4664 joss-reviews](https://github.com/openjournals/joss-reviews/issues/4664#issue-1334082133)

hi @luyuhao0326, I have some minor comments for the paper. Aside from that, I think the paper is written clearly and has all the key components required by JOSS.

-URL on line 77 needs to be checked.
-Line 42, 44, and 66 has minor typos.
