### 2022-09-09 UrbanRuralSystems/ur-scape#9 @winstonyym
JOSS Review - Functionality
[#4664 joss-reviews](https://github.com/openjournals/joss-reviews/issues/4664#issue-1334082133)

I am not able to test the planning and reachability analysis tools as they are not available for world data. The other tools work well. The ur-scape download page requires user credentials to access city level data. Would you be able to share for testing of functionality? I understand that some data might be private due to stakeholder relationships.
