### 2022-09-09 #4664 @winstonyym
I have completed a first pass of the software and tested the major functionalities. Overall, ur-scape is a beautifully designed software and has clear pragmatic implications for urban planning. It is easy to use and quite aesthetically pleasing. The focus on rapid urbanisation issues in Southeast Asia is indeed a critical and under-studied area. This is especially important given that more than 50% of urbanisation will happen in Asia over the next decade. Software would go a long way to helping planners and local decision makers in these cities. I have raised some minor issues for the authors to consider.
