+++ Paper/paper.md
@@ -63,7 +63,7 @@ ur-scape was developed in collaboration with a number of government agencies and
 
 ur-scape was developed in both game engine (Unity) and web-based environments. The desktop game-engine-enabled version supports real-time rendering. The desktop-based version compatible with a variety of platforms including Windows, MacOS, and Web is currently under development.
 
-Generally, ur-scape is structured under four key components: preparation, data, app, and develop (\autoref{fig: 3}). In the preparation stage, users interact with Mapbox to prepare the custom base maps they wish to render in the application <span>https://ur-scape.sec.sg/en/Tutorials/Create_a_Custom_Map_in_Mapbox</span>. In the current version of ur-scape (version 0.9.95), users also rely on QGIS to prepare and export both raster and vector data to a .csv format [https://ur-scape.sec.sg/en/Installation/QGIS_Plugin_Installation]. Metadata input is a required step when importing data into ur-scape. In the Data stage, data from the previous stage (i.e. csv files) are placed in ur-scape’s Data folder along with other data files (i.e. in-built tools and application configuration data files) used for the application’s interpretation. This leads to the App stage whereby the application interprets all the data files and renders them on an interactive map (set up via Mapbox during the initial installation of ur-scape). A list of sites containing data layers are also available for users to interact with. Site panel is a useful feature for projects involving local datasets with different data layer categories or in a different language.
+Generally, ur-scape is structured under four key components: preparation, data, app, and develop (\autoref{fig: 3}). In the preparation stage, users interact with Mapbox to prepare the custom base maps they wish to render in the application [Detailed tutorials](https://ur-scape.sec.sg/en/Tutorials/Create_a_Custom_Map_in_Mapbox). In the current version of ur-scape (version 0.9.95), users also rely on QGIS to prepare and export both raster and vector data to a .csv format [https://ur-scape.sec.sg/en/Installation/QGIS_Plugin_Installation]. Metadata input is a required step when importing data into ur-scape. In the Data stage, data from the previous stage (i.e. csv files) are placed in ur-scape’s Data folder along with other data files (i.e. in-built tools and application configuration data files) used for the application’s interpretation. This leads to the App stage whereby the application interprets all the data files and renders them on an interactive map (set up via Mapbox during the initial installation of ur-scape). A list of sites containing data layers are also available for users to interact with. Site panel is a useful feature for projects involving local datasets with different data layer categories or in a different language.
 
 Users can make use of the in-built tools to interact with and analyse the GIS data. Currently, ur-scape supports a total number of 11 tools with various complexity and data requirements [https://ur-scape.sec.sg/en/User_Guide/ur-scape/Tools_Panel]. The folder structure of existing and new tools, as well as the application’s platform configuration files are illustrated in \autoref{fig: 3}.
 
