### 2022-09-12 UrbanRuralSystems/ur-scape#8 @luyuhao0326
- line 77: the link is working in markdown. But in JOSS generated PDF preview, there is an additional "-" in the link and it seems to be causing the issue.
- line 66: changed Bundung to Bandung
- line 44: changed seemless to seamless
- line 42: changed ultising to utilising
