### 2022-09-12 UrbanRuralSystems/ur-scape#8 @winstonyym
Great, thanks for fixing these issues. I think this issue can be closed.
