### 2022-09-28 UrbanRuralSystems/ur-scape#9 @luyuhao0326
Hello @winstonyym, unfortunately due to the NDA, we can't share the local city data. However, we have an alternative open-source dataset for you to test the software and there is even a video to guide you and other users. You can access them freely via this [link](https://urstudio.sec.sg/courses/lesson-2-vectors/lesson/how-to-analyse-data-in-ur-scape/?page_tab=files). It might ask you to register for an account first. Let me know if you have any further questions :)
