### 2022-09-30 UrbanRuralSystems/ur-scape#9 @winstonyym
Hi @luyuhao0326, I totally understand. Thank you for the clear instructions. Overall, I managed to import the Cirebon dataset and tested the package functionalities. I also have to point out that the entire experience from watching the video to analysis was very clear. Snapshots are a really clever and convenient idea! I will proceed to close this issue.
