### 2022-09-30 #4664 @winstonyym
I have completed my second pass of ur-scape and expanding testing of the main package functionalities. The authors have kindly addressed the raised issues and we have resolved them successfully. The package works really well, use cases are interesting as well. It was quite an amazing user experience and ur-scape will definitely benefit many city planners in Southeast Asia.
