### 2022-10-05 #4664 @cforgaci
I have concluded my first review in which I tested the functionalities of ur-scape and examined the documentation. The software is a beautifully designed data visualization and analysis tool with a very accessible interface and considerable impact demonstrated with specific use cases. It would be great if the excellent video material and training data mentioned in https://github.com/UrbanRuralSystems/ur-scape/issues/9 would be made more accessible in the documentation. I raised a couple of minor issues that should be easy to resolve.
