+++ Paper/paper.md
@@ -30,11 +30,11 @@ bibliography: paper.bib
 
 The growing technical sophistication of city planning software [@Bettencourt:2010],[@Batty:2013] has produced increasingly specialised work flows that tend to alientate, rather than engage, stakeholders [@Batty:2021], [@Barns:2019], [@Yap:2022]. This contradicts the basic purpose of such software – or Planning Support Systems (PSS) – to improve the sustainability and long-term resilience of cities by harnessing increasing volumes of digital data to encourage greater stakeholder participation in city-making processes [@Maliene:2011], [@Flacke:2020], [@Geertman:2009], [@Lee:2014], [@Zwick:2010], [@Haddad:2012], [@vonRichthofen:2022]. 
 
-The disconnection of data from stakeholders is especially debilitating for rapidly urbanising regions in Asia and Africa. Cities in these regions are growing at unprecedented rates [@UN:2018] and typically mix both formal and informal, urban and rural land-uses. The uneven, dynamic and ambiguous settlement patterns that result are not always readily described with definitive, boundary-oriented and line-based graphical conventions. The policy makers, developers, investors, civil society actors, academics and the general public here urgently need access to reliable data in appropriate formats to support the planning of their cities and regions. 
+The disconnection of data from stakeholders is especially debilitating for rapidly urbanising regions in Asia and Africa. Cities in these regions are growing at unprecedented rates [@UN:2018] and typically mix both formal and informal, urban and rural land-uses. The uneven, dynamic and ambiguous settlement patterns and data landscape that result are not always readily described with definitive, boundary-oriented and line-based graphical conventions. The policy makers, developers, investors, civil society actors, academics and the general public here urgently need access to reliable data in appropriate formats to support the planning of their cities and regions. 
+
+ur-scape aims to offer solutions as part of the dataful way of city-making and city-managing. Specifically, ur-scape is an open-source planning support software designed to respond to the issues of rapid urbanisation in Asia and Africa. Supported by Unity game-engine, ur-scape aims to improve stakeholder access to data through an innovative data rendering format and an intuitive UI/UX concept.
 
 # Statement of need
-
-ur-scape is an open-source planning support software designed to respond to the challenges of rapid urbanisation in Asia and Africa. Supported by Unity game-engine, ur-scape aims to improve stakeholder access to data through an innovative data rendering format and an intuitive UI/UX concept.
 
 The data rendering format prioritises (raster-based) continuous grid-based data fields. We call this techniques ‘gridded Venn diagrams’ (\autoref{fig: 1}). ur-scape displays gridded data with an additive colour-mixed Venn diagrammatic representation at each centroid. Each coloured circle in the Venn diagram carries data from a given data set, be it population density, poverty, night-light, or CO2 emissions. As each centroid can accommodate a theoretically infinite number of colour-mixed circles arrayed around it, this rendering format facilitates layering and viewing multiple datasets at once. By avoiding the cumbersome data layering techniques typical of current PSS (through manipulations of opacity and transparency levels), this format encourages rapid and simple analyses of diverse kinds of data.
 
