### 2022-10-12 815202b
Fixed typos + some formatting
