+++ Paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Summary
 
-The growing technical sophistication of city planning software [@Bettencourt:2010],[@Batty:2013] has produced increasingly specialised work flows that tend to alientate, rather than engage, stakeholders [@Batty:2021], [@Barns:2019], [@Yap:2022]. This contradicts the basic purpose of such software – or Planning Support Systems (PSS) – to improve the sustainability and long-term resilience of cities by harnessing increasing volumes of digital data to encourage greater stakeholder participation in city-making processes [@Maliene:2011], [@Flacke:2020], [@Geertman:2009], [@Lee:2014], [@Zwick:2010], [@Haddad:2012], [@vonRichthofen:2022]. 
+The growing technical sophistication of city planning software [@Bettencourt:2010,@Batty:2013] has produced increasingly specialised work flows that tend to alineate, rather than engage, stakeholders [@Batty:2021,@Barns:2019,@Yap:2022]. This contradicts the basic purpose of such software -- or Planning Support Systems (PSS) -- to improve the sustainability and long-term resilience of cities by harnessing increasing volumes of digital data to encourage greater stakeholder participation in city-making processes [@Maliene:2011,@Flacke:2020,@Geertman:2009,@Lee:2014,@Zwick:2010,@Haddad:2012,@vonRichthofen:2022]. 
 
 The disconnection of data from stakeholders is especially debilitating for rapidly urbanising regions in Asia and Africa. Cities in these regions are growing at unprecedented rates [@UN:2018] and typically mix both formal and informal, urban and rural land-uses. The uneven, dynamic and ambiguous settlement patterns and data landscape that result are not always readily described with definitive, boundary-oriented and line-based graphical conventions. The policy makers, developers, investors, civil society actors, academics and the general public here urgently need access to reliable data in appropriate formats to support the planning of their cities and regions. 
 
@@ -40,7 +40,7 @@ The data rendering format prioritises (raster-based) continuous grid-based data 
 
 ![Illustrations of gridded Venn diagram concept at various scales.\label{fig: 1}](Figure_1.png)
 
-This approach is combined with an intuitive UI/UX concept based on simple and well-known gesture-based interaction formats – touch, pinch, zoom. The interaction concept is supported by a range of intuitive and "on-the-fly" analytical tools – such as contouring of multiple geospatial dataset intersections, transects and dissects – that make the most of the gridded Venn diagram approach in combination with multiple base maps (satellite, cadastral, traffic, and OSM). Utilising Unity’s processing and rendering capability, ur-scape is able to dynamically adjust image resolution based on the zoom level (i.e. extent of the dataset) to provide a fluent and seamless user experience even with large global datasets. Detailed descriptions and how-to instructions can be found in our website: [https://ur-scape.sec.sg/en/Tutorials] 
+This approach is combined with an intuitive UI/UX concept based on simple and well-known gesture-based interaction formats – touch, pinch, zoom. The interaction concept is supported by a range of intuitive and "on-the-fly" analytical tools -- such as contouring of multiple geospatial dataset intersections, transects and dissects -- that make the most of the gridded Venn diagram approach in combination with multiple base maps (satellite, cadastral, traffic, and OSM). Utilising Unity’s processing and rendering capability, ur-scape is able to dynamically adjust image resolution based on the zoom level (i.e. extent of the dataset) to provide a fluent and seamless user experience even with large global datasets. Detailed descriptions and how-to instructions can be found in our website: https://ur-scape.sec.sg/en/Tutorials
 
 This gives ur-scape a number of advantages over conventional planning support software:
 
@@ -63,15 +63,15 @@ ur-scape was developed in collaboration with a number of government agencies and
 
[…]
