### 2022-10-12 #4664 @hugoledoux
I see that the issues from the reviewers have been fixed (pretty quick!) and that the reviewers are happy.

One thing we need to discuss however, and it's a not-so-nice topic: authorship. I see that only 3 of the six authors have contributed to the code, and JOSS works differently from academic. We don't allow the person funding a project to put their names because of the fact that they have the money. The rules are there: https://joss.readthedocs.io/en/latest/submitting.html?highlight=authorship#authorship You'll see that it's not only git commits that count, but since I can't see teh contributions of 3 of the authors (especially the first one), could you please elaborate on their contributions to this package?

Also, I went over your paper and fixed a few things, please accept this PR: https://github.com/UrbanRuralSystems/ur-scape/pull/12
