### 2022-10-21 #4664 @luyuhao0326
Thank you @hugoledoux for coordinating our submission. Your question on (co)authorship is critical and I will clarify the roles of each author below.

As you may have noticed, ur-scape is relatively mature with its own UI/UX and a large (growing) body of users (and hopefully more developers with this submission at JOSS). One thing that I can be clear of is that none of the co-authors were included simply because of their "purely financial and organizational" roles in the ur-scape development.

In general, ur-scape is supported by three categories of contributors: designers, scientists, and software engineers. Some of them are coders on Github, and some of them are not. Some of them work astride multiple categories, and some of them have a very focused work task given their expertise. In our submission, we included key contributors from each category to credit their work in the development of ur-scape. In addition, due to the nature and key audience of JOSS, we prioritized contributors with a more software development focus.

The first three authors (Neudecker, Joos, and Zaol-kefli) are key contributors that translated ur-scape from a vision into usable software. Neudecker, in particular, with his expertise in UX/UI and excellent programming skills, not only created a smooth and intuitive user experience but also managed to help us overcome many technical obstacles (e.g. qgis importer).

Lu and Mangal provided support in data management, GIS, and capacity building. ur-scape was never intended to compete with ESRI and QGIS as a mapping software. It however heavily relies on principles and techniques that are crucial to the accuracy and precision of the software. Even though these technicalities are often hidden behind the curtains in ur-scape, we need experts like Lu and Mangel to ensure the science and relevance to the GIS community. Mangal with her background in urban designing also contributes to bringing forth the tool to a range of users varying from experts to non-experts through engagements such as Charrettes and collaborative projects with City governments. She has also developed the capacity-building platform for ur-scape to facilitate an easy uptake of the software. Lu also contributed to the writing, submission, and revision of this submission.

Cairns, the principal investigator of ur-scape, leads the team of software developers and scientists, as well as external designers and professionals to make sure that ur-scape is a piece of software that answers the needs of real-world planners and policy-makers. With his training in architecture, Cairns also pushes ur-scape in its pragmatic and aesthetic capabilities, two important qualities that made stakeholder engagement and idea exchange in ur-scape so effortless and pleasant. Cairns also contributed to the writing of this article.

Ur-scape has also welcomed other developers and contributors in the past, which we have acknowledged on its Github homepage. The authors listed above are the current core team of ur-scape.
