### 2022-10-21 #4664 @hugoledoux
thanks for the statement about authorship, this looks good.

Both reviewers recommend acceptance and your package fulfils all the criteria, at this point could you:

- [ ] Accept my PR https://github.com/UrbanRuralSystems/ur-scape/pull/12 -- you just close it w/o accepting it
