### 2022-10-31 #4664 @luyuhao0326
Hi @hugoledoux I have addressed the points from above. I am still learning GitHub so please let me know if I missed anything. Thanks again for your support!

- PR: accpepted & merged
- new release: v0.9.96 has been created and published
- software is archived under the ETH research collection reserved DOI (*pending approval): 10.3929/ethz-b-000578490
- ORCID for Yuhao Lu: 0000-0001-8574-1999
