### 2022-11-03 #4664 @luyuhao0326
@hugoledoux DOI of the archived software is now active via ETH Research Collection service:  [10.3929/ethz-b-000578490](https://doi.org/10.3929/ethz-b-000578490)
