### 2022-11-04 #4664 @hugoledoux
The archive doesn't have the correct metadata: only one author is there.

The list of authors should be exactly the same as that of the paper.
