### 2022-11-07 #4664 @luyuhao0326
@hugoledoux I have communicated with ETH Research Collection team and this issue should be ressovled by now. The record can be accessed [here](https://www.research-collection.ethz.ch/handle/20.500.11850/578490).
