### 2022-11-15 #4664 @hugoledoux
@openjournals/sbcs-eics I went too quickly, sorry. All the DOIs are wrong, the authors will fix those and then I'll contact you again. Sorry.

@winstonyym fix the DOIs as suggested above (https://github.com/openjournals/joss-reviews/issues/4664#issuecomment-1315329310), then we can continue with accepting the paper formally.
