+++ Paper/paper.bib
@@ -20,7 +20,7 @@
     Author = {{Yap}, W. and {Janssen}, P. and {Biljecki}, F},
     Journal = {Computers, Environment and Urban Systems},
     Title = {{Free and open source urbanism: Software for urban planning practice}},
-    Year = 2022
+    Year = 2022https://doi.org/10.1057/udi.2010.25
 }
 
 @report{ADB:2019,
@@ -69,7 +69,7 @@
 @article{Bettencourt:2010,
     Author = {{Bettencourt}, L and {Geoffrey}, W},
     Title = {{A unified theory of urban living}},
-    url = {https://www.nature.com/articles/467912a},
+    url = {https://doi.org/10.1038/467912a},
     Year = 2010
 }
 
