### 2022-11-17 #4664 @luyuhao0326
@hugoledoux I am not sure what to fix for the DOIs. I tested them out by copying and pasting the links to my browser and all the links took me to the article that I was citing. I did notice some of them were just a URL while others are DOIs. Could that be causing the issue?  Sorry if it is an amateur mistake.
