### 2022-11-17 #4664 @hugoledoux
so you don't store DOI but URL, and only the numbers need to be there, without the `http://doi.org/` part.

As an example, [this line](https://github.com/UrbanRuralSystems/ur-scape/blob/paper/Paper/paper.bib#L19) should be:
```
doi={10.1016/j.compenvurbsys.2022.101825},
```
