+++ Paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{Batty:2021,
-    url = {https://doi.org/10.1177/23998083211016122},
+    url = {10.1177/23998083211016122},
     Author = {{Batty}, M.},
     Title = {{The digital transformation of planning}},
     Journal = {Environment and Planning B: Urban Analytics and City Science},
@@ -16,7 +16,7 @@
 }
 
 @article{Yap:2022,
-    url = {https://doi.org/10.1016/j.compenvurbsys.2022.101825},
+    url = {10.1016/j.compenvurbsys.2022.101825},
     Author = {{Yap}, W. and {Janssen}, P. and {Biljecki}, F},
     Journal = {Computers, Environment and Urban Systems},
     Title = {{Free and open source urbanism: Software for urban planning practice}},
@@ -53,7 +53,7 @@
 }
 
 @article{Maliene:2011,
-    url = {https://doi.org/10.1057/udi.2010.25},
+    url = {10.1057/udi.2010.25},
     Author = {{Maliene}, V. and {Grigonis}, V. and {Palevičius}, V and {Griffiths}, S},
     Title = {{Geographic information system: Old principles with new capabilities}},
     Year = 2011
@@ -69,7 +69,7 @@
 @article{Bettencourt:2010,
     Author = {{Bettencourt}, L and {Geoffrey}, W},
     Title = {{A unified theory of urban living}},
-    url = {https://doi.org/10.1038/467912a},
+    url = {10.1038/467912a},
     Year = 2010
 }
 
@@ -94,7 +94,7 @@
 }
 
 @article{Flacke:2020,
-    url = {https://doi.org/10.3390/ijgi9010049},
+    url = {10.3390/ijgi9010049},
     Author = {{Flacke} J, and {Rehana}, S, and {Aguilar}, R},
     Title = {{Strengthening Participation Using Interactive Planning Support Systems: A Systematic Review}},
     Year = 2020
@@ -134,7 +134,7 @@
 }
 
 @article{Gilbert:2015,
-    url = {https://doi.org/10.1038/sdata.2017.1},
+    url = {10.1038/sdata.2017.1},
     Author = {{Gilbert}, Laporte and {Stefan}, Nickel and {Francisco}, Saldanha da Gama},
     Title = {{High Resolution Global Gridded Data for Use in Population Studies}},
     Year = 2015
@@ -166,7 +166,7 @@
 }
 
 @journal{vonRichthofen:2022,
-    url = {https://doi.org/10.1177/08854122211068526},
+    url = {10.1177/08854122211068526},
     Author = {{von Richthofen}, A and {Herthogs}, P and {Kraft}, M and {Cairns}, S},
     Title = {{Semantic City Planning Systems (SCPS): A Literature Review}},
     Year = 2022
@@ -179,7 +179,7 @@
 }
 
 @article{Andrew:2017,
-    url = {https://doi.org/10.1038/sdata.2017.4},
+    url = {10.1038/sdata.2017.4},
     Author = {{Andrew}, J Tatem},
     Title = {{WorldPop, Open Data for Spatial Demography}},
     Year = 2017
