+++ Paper/paper.bib
@@ -20,7 +20,7 @@
     Author = {{Yap}, W. and {Janssen}, P. and {Biljecki}, F},
     Journal = {Computers, Environment and Urban Systems},
     Title = {{Free and open source urbanism: Software for urban planning practice}},
-    Year = 2022https://doi.org/10.1057/udi.2010.25
+    Year = 2022
 }
 
 @report{ADB:2019,
