+++ Paper/paper.bib
@@ -109,15 +109,15 @@
 
 @article{Geertman:2013,
     Author = {{Geertman}, Stan and {Toppen}, F. and {Stillwell}, J.},
-    Title = {{Planning Support Systems for Sustainable Urban Development}},
-    url = {https://link.springer.com/book/10.1007/978-3-642-37533-0?noAccess=true},
+    Title = {{Introduction to Planning Support Systems for Sustainable Urban Development}},
+    doi = {10.1007/978-3-642-37533-0?noAccess=true},
     Year = 2013
 }
 
 @article{Geertman:2015,
     Author = {{Geertman}, Stan and {Ferreira}, J. and {Goodspeed}, R. and {Stillwell}, John},
     Title = {{Introduction to ‘planning support systems and smart cities’}},
-    url = {https://link.springer.com/chapter/10.1007/978-3-319-18368-8_1},
+    doi = {10.1007/978-3-319-18368-8_1},
     Year = 2015
 }
 
@@ -153,7 +153,7 @@
 }
 
 @article{Pelzer:2013,
-    url = {https://link.springer.com/chapter/10.1007/978-3-319-08299-8_21},
+    doi = {10.1007/978-3-319-08299-8_21},
     Author = {{Pelzer}, P. and {Brömmelstroet}, M. T. and {Geertman} S.},
     Title = {{Geodesign in Practice: What about the urban designers}},
     Year = 2013
