+++ Paper/paper.bib
@@ -129,6 +129,7 @@
 
 @article{Kelly:2002,
     Author = {{Kelly}, Philip F},
+    doi = {10.1111/1475-5661.00062},
     Title = {{Spaces of labour control: Comparative perspectives from Southeast Asia}},
     Year = 2002
 }
@@ -161,6 +162,7 @@
 
 @article{Christopher:2006,
     Author = {{Christopher}, J Pettit and {William}, Cartwright and {Michael}, Berry},
+    doi = {10.2104/ag060022},
     Title = {{Geographical visualization: A participatory planning support tool for imagining landscape futures}},
     Year = 2006
 }
@@ -174,6 +176,7 @@
 
 @article{Haddad:2012,
     Author = {{Haddad}, M},
+    doi = {10.1177/0739456X15581606},
     Title = {{A framework for geodesign: Changing geography by design}},
     Year = 2012
 }
