+++ Paper/paper.bib
@@ -8,7 +8,6 @@
 }
 
 @book{Barns:2019,
-    url = {https://link.springer.com/book/10.1007/978-981-32-9725-8?noAccess=true},
     Author = {{Barns}, S},
     Title = {{Platform urbanism: negotiating platform ecosystems in connected cities}},
     Publisher = {Springer Nature},
