+++ Paper/paper.bib
@@ -203,5 +203,5 @@
     Author = {{Zwick}, P.},
     Title = {{The world beyond GIS}},
     url = {https://trid.trb.org/view/925657},
-    Year = 2010
+    Year = 2010 
 } 
