+++ Paper/paper.bib
@@ -72,10 +72,11 @@
     Year = 2010
 }
 
-@article{Brail:2008,
-    Author = {{Brail}, Richard K},
+@article{Kaza:2009,
+    doi = {10.1080/01944360903409493},
+    Author = {{Kaza}, Nikhil},
     Title = {{Planning support systems for cities and regions}},
-    Year = 2008
+    Year = 2009
 }
 
 @book{Lee:2014,
@@ -109,7 +110,7 @@
 @article{Geertman:2013,
     Author = {{Geertman}, Stan and {Toppen}, F. and {Stillwell}, J.},
     Title = {{Introduction to Planning Support Systems for Sustainable Urban Development}},
-    doi = {10.1007/978-3-642-37533-0?noAccess=true},
+    doi = {10.1007/978-3-642-37533-0},
     Year = 2013
 }
 
