### 2022-11-18 #4664 @hugoledoux
@openjournals/sbcs-eics the DOIs have been fixed and thus the formal acceptance can be done. I imagine this doesn't necessitate a new DOI/version since it was just removing the `http://doi.org/` from the `.bib`? If it does, let us know here.
