### 2022-11-20 #4664 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
