+++ Paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Summary
 
-The growing technical sophistication of city planning software [@Bettencourt:2010,@Batty:2013] has produced increasingly specialised workflows that tend to alienate, rather than engage stakeholders [@Batty:2021,@Barns:2019,@Yap:2022]. This contradicts the basic purpose of such software -- or Planning Support Systems (PSS) -- to improve the sustainability and long-term resilience of cities by harnessing increasing volumes of digital data to encourage greater stakeholder participation in city-making processes [@Maliene:2011,@Flacke:2020,@Geertman:2009,@Lee:2014,@Zwick:2010,@Haddad:2012,@vonRichthofen:2022]. 
+The growing technical sophistication of city planning software [@Bettencourt:2010;@Batty:2013] has produced increasingly specialised workflows that tend to alienate, rather than engage stakeholders [@Batty:2021;@Barns:2019;@Yap:2022]. This contradicts the basic purpose of such software -- or Planning Support Systems (PSS) -- to improve the sustainability and long-term resilience of cities by harnessing increasing volumes of digital data to encourage greater stakeholder participation in city-making processes [@Maliene:2011;@Flacke:2020;@Geertman:2009;@Lee:2014;@Zwick:2010;@Haddad:2012;@vonRichthofen:2022]. 
 
 The disconnect of data from stakeholders is especially debilitating for rapidly urbanising regions in Asia and Africa. Cities in these regions are growing at unprecedented rates [@UN:2018] and typically mix both formal and informal, urban and rural land-uses. The uneven, dynamic and ambiguous settlement patterns and data landscape that result are not always readily described with definitive, boundary-oriented and line-based graphical conventions. The policy makers, developers, investors, civil society actors, academics and the general public urgently need access to reliable data in appropriate formats to support the planning of their cities and regions. 
 
