### 2022-11-20 #4664 @luyuhao0326
@arfon PR merged! And thanks for checking this for us. I should have done this myself first. Overall, this has been a very enjoyable and productive process with JOSS. Look forward to engaging with you, @hugoledoux, and other editors in the coming years.
